updated redux dependencies for ui samples migrated to new redux style api for pattern-list-details-redux
