+++ ui/all-samples/package.json
@@ -39,10 +39,10 @@
     "query-string": "^6.7.0",
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
-    "react-redux": "^7.2.0",
+    "react-redux": "^8.0.1",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.2.0",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
     "web-animations-js": "^2.3.2"
   }
 }
