+++ ui/pattern-analytics/src/index.js
@@ -1,6 +1,6 @@
 import {configure} from '@enact/analytics';
 import {forKeyCode} from '@enact/core/handle';
-import {createRoot} from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import 'web-animations-js';
 
 import App from './App';
