--- ui/pattern-list-details-redux/src/actions/actions.js
@@ -1,14 +1,0 @@
-export const changeCountry = (country) => ({
-	type: 'CHANGE_COUNTRY',
-	country
-});
-
-export const changeCity = (city) => ({
-	type: 'CHANGE_CITY',
-	city
-});
-
-export const changeZoom = (zoom) => ({
-	type: 'ZOOM',
-	zoom
-});
