--- ui/pattern-list-details-redux/src/reducers/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "rootReducer.js"
-}
