+++ ui/pattern-ls2request/package.json
@@ -39,8 +39,8 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-redux": "^7.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
     "web-animations-js": "^2.3.2"
   }
 }
