+++ ui/pattern-ls2request/src/index.js
@@ -1,4 +1,4 @@
-import {createRoot} from 'react-dom';
+import {createRoot} from 'react-dom/client';
 import 'web-animations-js';
 
 import App from './main';
